more clean up and dir restructuring with back button re-routing
