+++ src/routes/index.tsx
@@ -1,10 +1,7 @@
-import Hero from "@/components/custom/Hero";
+import Hero from "@/pages/home/Hero";
 import NotFound from "@/pages/not_found/NotFound";
 import { RouteConfig } from "@/types/RouteConfig";
-import About from "@/pages/about/about/About";
-// import Shop from "@/pages/shop/Shop";
-// import Contact from "@/pages/contact/Contact";
-// import Gallery from "@/pages/gallery/Gallery";
+import About from "@/pages/about/About";
 
 export const routes: RouteConfig[] = [
   {
@@ -15,18 +12,6 @@ export const routes: RouteConfig[] = [
     path: "/about",
     element: <About />,
   },
-  // {
-  //   path: "/shop",
-  //   element: <Shop />,
-  // },
-  // {
-  //   path: "/contact",
-  //   element: <Contact />,
-  // },
-  // {
-  //   path: "/gallery/all",
-  //   element: <Gallery />,
-  // },
   {
     path: "*",
     element: <NotFound />,
